pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cacherator"
version = "1.2.2"
description = "Persistent JSON caching for Python with async support - cache function results and object state effortlessly."
readme = "README.md"
authors = [
    { name = "Arved Klöhn", email = "arved.kloehn@gmail.com" }
]
license = "MIT"
keywords = [
    "cache",
    "caching",
    "json",
    "persistent",
    "async",
    "decorator",
    "memoization",
    "storage",
    "dynamodb",
    "aws"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
    "Framework :: AsyncIO"
]
requires-python = ">=3.7"

dependencies = [
    "python-slugify>=8.0.0",
    "logorator>=1.0.0",
    "dynamorator>=0.1.5",
]

[project.optional-dependencies]
dynamodb = ["boto3>=1.26.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "build>=0.10.0",
    "twine>=4.0.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
]

[project.urls]
Source = "https://github.com/Redundando/cacherator"
Issues = "https://github.com/Redundando/cacherator/issues"
Documentation = "https://github.com/Redundando/cacherator#readme"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "*.egg-info", "dist", "build", "cache", "data", "test_cache", ".venv", "examples"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
